+++ crates/fluaoslib/src/lib.rs
@@ -2,4 +2,5 @@
 // a library Create for flua which contains some OS depending Code
 // and other Stuff
 
+pub mod folderdialog;
 pub mod windows_utf8;