+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - added registering the SQLite API because i forgot it
 - added Logger Module and changed some stuff in the config
 - logger can now be used directly from Lua
+- add Folder and File Dialog Functions to the Lue OS Lib
 
 ## 0.2.0 - 20.10.2025